refactor: replace align_seg.py and align_points.py with neuroreg equivalents
- Delete recon_surf/align_seg.py -- fully superseded by neuroreg.cli.segreg
  (neuroreg.cli.segreg already used in talairach-reg.sh)
- Delete recon_surf/align_points.py -- superseded by neuroreg.segreg.points
- recon_surf/rotate_sphere.py: import find_rotation from neuroreg.segreg.points;
  inline rmat2angles (FreeSurfer-specific, not in neuroreg) as a local helper
- CorpusCallosum/registration/midsagittal_plane_alignment.py: import find_rigid
  from neuroreg.segreg.points instead of recon_surf.align_points
